Добавить разделение проекта на модули
