+++ public/index.html
@@ -453,7 +453,6 @@
         </form>
       </article>
     </template>
-
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
