+++ src/main.js
@@ -1,163 +1,102 @@
-(function () {
-  const filterTemlate = document.querySelector(`#filter-template`);
+import {randomInteger, randomArrayItem, removeChilds} from './utils.js';
+import createFilterElement from './filter.js';
+import createTaskElement from './task.js';
 
-  /**
-   * Возвращает элемент фильтра, созданный из шаблона #filter-template
-   * @param {Object} filterDefinition - описание фильтра
-   * @return {Node} - возвращает элемент фильтра
-   */
-  function createFilterElement(filterDefinition) {
-    const filterElement = filterTemlate.content.cloneNode(true);
+/**
+ * Функция отображает список фильтров
+ */
+function renderFilters() {
+  const filterDefinitions = [
+    {
+      id: `all`,
+      name: `ALL`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: true
+    },
 
-    const filterInputElement = filterElement.querySelector(`input`);
-    filterInputElement.setAttribute(`id`, `filter__${filterDefinition.id}`);
-    filterInputElement.checked = filterDefinition.isActive;
-    filterInputElement.disabled = filterDefinition.tasksCount === 0;
+    {
+      id: `overdue`,
+      name: `OVERDUE`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: false
+    },
 
-    const filterLabelElement = filterElement.querySelector(`label`);
-    filterLabelElement.innerHTML = `${filterDefinition.name} <span class="filter__${filterDefinition.id}-count">${filterDefinition.tasksCount}</span>`;
-    filterLabelElement.setAttribute(`for`, `filter__${filterDefinition.id}`);
-    filterLabelElement.addEventListener(`click`, filterDefinition.onClick);
+    {
+      id: `today`,
+      name: `TODAY`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: false
+    },
 
-    return filterElement;
-  }
+    {
+      id: `favorites`,
+      name: `FAVORITES`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: false
+    },
 
-  /**
-   * Функция возвращает целое случайное число в диапазоне [0, max)
-   * @param {*} max - максимальное число ()
-   * @return {Node}
-   */
-  function randomInteger(max) {
-    return Math.floor(Math.random() * max);
-  }
+    {
+      id: `repeating`,
+      name: `REPEATING`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: false
+    },
 
-  /**
-   * Функция отображает список фильтров
-   */
-  function renderFilters() {
-    const filterDefinitions = [
-      {
-        id: `all`,
-        name: `ALL`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: true
-      },
+    {
+      id: `tags`,
+      name: `TAGS`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: false
+    },
 
-      {
-        id: `overdue`,
-        name: `OVERDUE`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: false
-      },
-
-      {
-        id: `today`,
-        name: `TODAY`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: false
-      },
-
-      {
-        id: `favorites`,
-        name: `FAVORITES`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: false
-      },
-
-      {
-        id: `repeating`,
-        name: `REPEATING`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: false
-      },
-
-      {
-        id: `tags`,
-        name: `TAGS`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: false
-      },
-
-      {
-        id: `archive`,
-        name: `ARCHIVE`,
-        tasksCount: randomInteger(10),
-        onClick: showTasks,
-        isActive: false
-      }
-    ];
-
-    const filtersFragment = document.createDocumentFragment();
-    const filterElements = filterDefinitions.map(createFilterElement);
-
-    for (const filterElement of filterElements) {
-      filtersFragment.appendChild(filterElement);
+    {
+      id: `archive`,
+      name: `ARCHIVE`,
+      tasksCount: randomInteger(10),
+      onClick: showTasks,
+      isActive: false
     }
+  ];
 
-    const filtersContainerElement = document.querySelector(`.main__filter`);
-    filtersContainerElement.appendChild(filtersFragment);
+  const filtersFragment = document.createDocumentFragment();
+  const filterElements = filterDefinitions.map(createFilterElement);
+
+  for (const filterElement of filterElements) {
+    filtersFragment.appendChild(filterElement);
   }
 
-  renderFilters();
+  const filtersContainerElement = document.querySelector(`.main__filter`);
+  filtersContainerElement.appendChild(filtersFragment);
+}
 
-  const taskTemplate = document.querySelector(`#card-template`);
+renderFilters();
 
-  /**
-   * Возвращает карточку задачи, созданную из шаблона #card-template
-   * @param {Object} taskDefinition - описание задачи
-   * @return {Node} - возвращает элемент карточки задачи
-   */
-  function createTaskElement(taskDefinition) {
-    const taskElement = taskTemplate.content.querySelector(`.card`).cloneNode(true);
-    taskElement.classList.add(`card--${taskDefinition.color}`);
-    return taskElement;
+const tasksContainerElement = document.querySelector(`.board__tasks`);
+
+/**
+ * Функция отображает случайное количество карточек задач
+ */
+function showTasks() {
+  removeChilds(tasksContainerElement);
+  const tasksCount = randomInteger(10);
+  const TASK_COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
+  const tasksDefinitions = Array(tasksCount).fill().map(() => (
+    {color: randomArrayItem(TASK_COLORS)})
+  );
+
+  const tasksFragment = document.createDocumentFragment();
+  for (const taskDefinition of tasksDefinitions) {
+    tasksFragment.appendChild(createTaskElement(taskDefinition));
   }
 
-  /**
-   * Функция удаляет все дочерние элементы
-   * @param {*} element - DOM элемент, из которого будут удалены дочерние
-   */
-  function removeChilds(element) {
-    while (element.firstChild) {
-      element.removeChild(element.firstChild);
-    }
-  }
+  tasksContainerElement.appendChild(tasksFragment);
+}
 
-  /**
-   * Функция возвращает случайный элемент массива
-   * @param {Array} array - массив
-   * @return {*} - возвращает случайный элемент массива
-   */
-  function randomArrayItem(array) {
-    return array[randomInteger(array.length)];
-  }
+showTasks();
 
-  const tasksContainerElement = document.querySelector(`.board__tasks`);
-
-  /**
-   * Функция отображает случайное количество карточек задач
-   */
-  function showTasks() {
-    removeChilds(tasksContainerElement);
-    const tasksCount = randomInteger(10);
-    const TASK_COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
-    const tasksDefinitions = Array(tasksCount).fill().map(() => (
-      {color: randomArrayItem(TASK_COLORS)})
-    );
-
-    const tasksFragment = document.createDocumentFragment();
-    for (const taskDefinition of tasksDefinitions) {
-      tasksFragment.appendChild(createTaskElement(taskDefinition));
-    }
-
-    tasksContainerElement.appendChild(tasksFragment);
-  }
-
-  showTasks();
-})();
