+++ webpack.config.js
@@ -1,6 +1,7 @@
 /* eslint-disable */
 
 const path = require(`path`);
+
 module.exports = {
   mode: `development`, // Режим сборки
   entry: `./src/main.js`, // Точка входа приложения
